fix(host): Stop unnecessarily trimming email bodies
The body was trimmed since it kept adding blank lines to the end of my
emails. I thought it was a Neovim fileformat=dos issue but it turned out
the culprit was an writeln!() by mistake when writing temporary files.

Hence replacing the writeln!() with write!() and stopping trimming
bodies.