fix(ext): Use correct identity to reply
Note this needs accountsRead permission to see MessageHeader.folder and
call accounts API.

Previously it always replied with the default identity, even if the
email was actually sent to a different account (and this account was
CC'ed). This is different from Thunderbird's built-in reply function
from UI.

Now we use the first identity of the account that's associated to the
email's folder. Considering most email providers don't allow
multi-identities, this should work most of the time.