feat: Reset compose state if error is recoverable
Previously if anything went wrong in the messaging host, we only showed
a notification to the user and the user was stuck with a greyed out
ExtEditorR button.

However in some cases, the error may be fixed by updating ExtEditorR's
and/or the editor's configurations, e.g.

- Host cannot create the temporary EML file
- Host cannot write to the temporary EML file
- Host cannot start the editor

Now when such an error arises, we re-enable ExtEditorR's button and
clear out all received messages for the compose window (which should be
empty in the first place).

For other errors, we keep the button greyed out to avoid overwriting any
data. And there are also errors like version mismatch where users have
to reload ExtEditorR.